requirements.txt
python-binance
pathway
pandas
numpy
streamlit
plotly
python-dotenv
statsmodels
matplotlib
streamlit-autorefresh